added flex-col into layout
