--- apps/web/app/[lng]/shop-list/page.tsx
@@ -1,12 +1,0 @@
-import Link from 'next/link';
-import { useTranslation } from '../../i18n';
-
-export default async function Page({ params: { lng } }) {
-  const { t } = await useTranslation(lng, 'shop-list');
-  return (
-    <>
-      <h1>{t('title')}</h1>
-      <Link href={`/${lng}`}>{t('back-to-home')}</Link>
-    </>
-  );
-}
