+++ apps/web/app/i18n/locales/en/translation.json
@@ -1,5 +1,5 @@
 {
   "title": "Hi there!",
-  "to-shop-list": "To second page",
+  "to-shops-page": "To second page",
   "to-map-page": "To client page"
 }
