+++ apps/web/app/i18n/locales/zh-CN/translation.json
@@ -1,5 +1,5 @@
 {
   "title": "你好",
-  "to-shop-list": "商家列表",
+  "to-shops-page": "商家列表",
   "to-map-page": "前往地图"
 }
